--- src/Controller/RegistrationController.php
@@ -1,91 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\Entity\User;
-use App\Form\RegistrationType;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
-
-class RegistrationController extends AbstractController
-{
-    /**
-     * @Route("/registration", name="registration")
-     * @param UserPasswordEncoderInterface $passwordEncoder
-     * @param EntityManagerInterface $em
-     * @param Request $request
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse|Response
-     */
-    public function registration(UserPasswordEncoderInterface $passwordEncoder, EntityManagerInterface $em, Request $request)
-    {
-        $user = new User();
-        dump($user);
-        $form = $this->createForm(RegistrationType::class, $user);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            // Etape de plus : hasher le mot de pass
-
-            $hash = $passwordEncoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($hash);
-            $user->setPseudo($user->getEmail());
-
-            $user->setRoles(['ROLE_USER']);
-
-
-            $em->persist($user);
-            $em->flush();
-
-            $this->addFlash("success", "Inscription OK !");
-
-            return $this->redirectToRoute('app_login');
-        }
-
-        return $this->render('registration/registration.html.twig', [
-            "form" => $form->createView()
-        ]);
-    }
-
-    /**
-     * @Route("/login", name="app_login")
-     * @param AuthenticationUtils $authenticationUtils
-     * @return Response
-     */
-    public function login(AuthenticationUtils $authenticationUtils): Response
-    {
-        // if ($this->getUser()) {
-        //     return $this->redirectToRoute('target_path');
-        // }
-
-        // get the login error if there is one
-        $error = $authenticationUtils->getLastAuthenticationError();
-        // last username entered by the user
-        $lastUsername = $authenticationUtils->getLastUsername();
-
-        return $this->render('registration/login.html.twig', ['last_username' => $lastUsername, 'error' => $error]);
-    }
-
-    /**
-     * @Route("/logout", name="app_logout")
-     */
-    public function logout()
-    {
-        throw new \Exception('This method can be blank - it will be intercepted by the logout key on your firewall');
-    }
-
-    /**
-     * @Route("/test", name="test")
-     */
-    public function test()
-    {
-        return $this->render('registration/test.html.twig', [
-            "form" => "form"
-        ]);
-    }
-
-
-}
