+++ src/Controller/SecurityController.php
@@ -2,13 +2,53 @@
 
 namespace App\Controller;
 
+use App\Entity\User;
+use App\Form\RegistrationType;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 
 class SecurityController extends AbstractController
 {
+    /**
+     * @Route("/registration", name="registration")
+     * @param UserPasswordEncoderInterface $passwordEncoder
+     * @param EntityManagerInterface $em
+     * @param Request $request
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|Response
+     */
+    public function registration(UserPasswordEncoderInterface $passwordEncoder, EntityManagerInterface $em, Request $request)
+    {
+        $user = new User();
+        dump($user);
+        $form = $this->createForm(RegistrationType::class, $user);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Etape de plus : hasher le mot de pass
+
+            $hash = $passwordEncoder->encodePassword($user, $user->getPassword());
+            $user->setPassword($hash);
+            $user->setPseudo($user->getEmail());
+
+            $user->setRoles(['ROLE_USER']);
+
+
+            $em->persist($user);
+            $em->flush();
+
+            $this->addFlash("success", "Inscription OK !");
+
+            return $this->redirectToRoute('app_login');
+        }
+
+        return $this->render('security/registration.html.twig', [
+            "form" => $form->createView()
+        ]);
+    }
     /**
      * @Route("/login", name="app_login")
      */
@@ -28,9 +68,20 @@ class SecurityController extends AbstractController
 
     /**
      * @Route("/logout", name="app_logout")
+     * @throws \Exception
      */
     public function logout()
     {
         throw new \Exception('This method can be blank - it will be intercepted by the logout key on your firewall');
+    }
+
+    /**
+     * @Route("/test", name="test")
+     */
+    public function test()
+    {
+        return $this->render('security/test.html.twig', [
+            "form" => "form"
+        ]);
     }
 }
