+++ templates/security/login.html.twig
@@ -3,26 +3,56 @@
 {% block title %}Log in!{% endblock %}
 
 {% block body %}
-<form method="post">
-    {% if error %}
-        <div class="alert alert-danger">{{ error.messageKey|trans(error.messageData, 'security') }}</div>
-    {% endif %}
+    <div class="container">
+        <div class="row">
+            <div class="col-sm-4 col-sm-4 offset-md-4">
+                <div class="card">
+                    <article class="card-body">
+                        <a href="{{ path('registration') }}" class="float-right btn btn-outline-primary">Sign up</a>
+                        <h4 class="card-title mb-4 mt-1">Connexion</h4>
+                        <hr>
+                        <form method="post">
+                            {% if error %}
+                                <div class="alert alert-danger">{{ error.messageKey|trans(error.messageData, 'security') }}</div>
+                            {% endif %}
 
-    {% if app.user %}
-        <div class="mb-3">
-            You are logged in as {{ app.user.username }}, <a href="{{ path('app_logout') }}">Logout</a>
+                            {% if app.user %}
+                                <div class="mb-3">
+                                    You are logged in as {{ app.user.username }}, <a href="{{ path('app_logout') }}">Logout</a>
+                                </div>
+                            {% endif %}
+                            <div class="form-group">
+                                <input  autofocus id="inputEmail" name="email" value="{{ last_username }}" class="form-control" placeholder="Email or login" type="email" required>
+                            </div> <!-- form-group// -->
+                            <div class="form-group">
+                                <input name="password" id="inputPassword" class="form-control" placeholder="******" type="password" required>
+                                <input type="hidden" name="_csrf_token"
+                                       value="{{ csrf_token('authenticate') }}"
+                                >
+                            </div> <!-- form-group// -->
+                            <div class="row">
+                                <div class="col-md-6">
+                                    <div class="form-group">
+                                        <button type="submit" class="btn btn-primary btn-block"> Login  </button>
+                                    </div> <!-- form-group// -->
+                                </div>
+
+                                <div class="col-md-6 text-right">
+                                    <a class="small" href="#">Forgot password?</a>
+                                </div>
+                            </div>
+                        </form>
+                        <div class="col-md-6">
+                            <button class="btn btn-block btn-primary" href="{{ path('app_logout') }}"> Deconnexion</button>
+                        </div>
+                    </article>
+                </div>
+            </div>
         </div>
-    {% endif %}
+    </div>
 
-    <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
-    <label for="inputEmail">Email</label>
-    <input type="email" value="{{ last_username }}" name="email" id="inputEmail" class="form-control" required autofocus>
-    <label for="inputPassword">Password</label>
-    <input type="password" name="password" id="inputPassword" class="form-control" required>
 
-    <input type="hidden" name="_csrf_token"
-           value="{{ csrf_token('authenticate') }}"
-    >
+
 
     {#
         Uncomment this section and add a remember_me option below your firewall to activate remember me functionality.
@@ -35,8 +65,5 @@
         </div>
     #}
 
-    <button class="btn btn-lg btn-primary" type="submit">
-        Sign in
-    </button>
-</form>
 {% endblock %}
+
